Added state machine to speculative objects.  A derived speculative can be in one of three states:
RUNNING --- This means that the speculative object is not the root
object and all the objects that it is speculating on have neither
completed nor aborted.

SUCCEEDED --- The speculative object is not the root object and all
the objects that it had speculated on have succeeded in committing.
The root object is now handling logic for readers and writers on dirty
value and this object should just forward commit and backout requests
to root object.

FAILED --- One of the objects we speculated on failed.  Should back
everything out.