Had a problem with previous version of SpeculativeObjects.  Their get_vals and set_vals held locks when they shouldn't have.  Ie, previous code looked like:
get_val
   lock
      super_get_val
   unlock

because super_get_val blocks, would just indefinitely hold onto
object's lock.  The ugly solution I came up to for this was to change
the interface that AtomicObjects provided so that they could take in a
ReentrantLock that they would unlock after they registered the
request, etc., but before they blocked on waiting for a response.