If the call is just a specualte call (not specualte container internals), then we speculate on just the wrapper of an object.  For instance, for the following code
   List(element: Number) some_list;
   List(element: Number) some_other_list;
       ...
   some_list = some_other_list;
   speculate(some_list);

Allows other events to now read/write on top of some_other_list when
accessing some_list.

However,

   List(element: Number) some_list;
   some_list.append(1);
   speculate(some_list);

does not allow other events to perform basic operations.  Eg.,

   some_list.size();

This is because speculate allows us to speculate on the pointer
wrapping the internal list, we cannot speculate on the contents of
that internal list.

To speculate on the internal value added a new method:

   speculate_container_internals(some_list);

This method should get the internal container pointed at and
call speculate on it, instead of the external wrapper.