Testing speculatives further.  If have a chain of speculated objects:
a -> b -> c

where a is the root speculative object, b derives from a's value, and c
derives from b's value, then want to ensure that if a is not backed out, but
b is then, b gets removed from chain.

Currently does not pass; it just hangs.  Next series of commits should fix.