If have a chain of speculateds:
a -> b -> c

where a is root.  If b is invalidated and b had been a write, now b
asks a to invalidate c.

Note 1: There seems to still be a bug with this logic.  Test case
MiddleInterruptedSpeculater does not pass.

Note 2: Incorrectly handling case where just has one reader.  In those
cases, should remove b and move c to a.