Previously had a bug in emitter for enums.  Had a condition that if a dot ast node had an enum type, would treat expression as a static access on enum declaration.  For instance, for
enum Day
{ MONDAY, TUESDAY, ...}

any x.y that was an enum would be emitted as Day.<value of y>.  This
was problematic because it could interfere with a struct with an enum
field, eg. for:

Struct Date
{
    enum Day d;
    ...
}

any access to d would be emitted as Day.d, instead of the struct
instance.d.  Ie.,

Struct Date date;
date.d;

would emit:
Day.d
instead of
date.d.

Fixed.