I'm going to reuse the java lib testing framework to test version message serialization and deserialization.
However, the Java library tests had fallen badly out of date and don't
even compile.  This commit updated all tests so that they compile.  It
appears that PartnersNoConflict test and PartnersModifyRemoteState
tests both fail.  This might be because of deprecation of reference
variables.  Either way, not focusing on this issue now.